refactor(desktop): read a knowledge-selection answer in one place
`readPrimary` and `readHidden` move out of `KnowledgeContext` into a module of
their own, `knowledge/knowledgeSelection.ts`, unchanged. The `/` palette and the
create-workflow modal each re-spelled the same two reads inline
(`primary_kb ?? active_kb`, `hidden_kbs ?? []`), and the next commits give them
a shared reader, which needs these two in a place all three callers can import.

No behavior change.
